# backend/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Mimi Platform API
  description: |
    REST API платформы Mimi — проектный менеджер с задачами, документами, канвасами и командами.

    **Аутентификация**: большинство эндпоинтов требуют `Authorization: Bearer <access_token>`.
    Получите токен через `POST /api/auth/login`.

    **Архитектура**: 9 независимых микросервисов, доступных через NGINX gateway на порту 8080.
  version: "1.0.0"
  contact:
    name: Mimi Dev Team
servers:
  - url: http://localhost:8080
    description: Локальный Docker Compose
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Team:
      type: object
      properties:
        id:        { type: string, format: uuid }
        name:      { type: string }
        ownerId:   { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
    RolePermissions:
      type: object
      properties:
        canViewDocs:   { type: boolean }
        canEditDocs:   { type: boolean }
        canViewTasks:  { type: boolean }
        canEditTasks:  { type: boolean }
        canViewBoards: { type: boolean }
        canEditBoards: { type: boolean }
        canManageTeam: { type: boolean }
    Role:
      type: object
      properties:
        id:          { type: string, format: uuid }
        teamId:      { type: string, format: uuid }
        name:        { type: string }
        permissions: { $ref: '#/components/schemas/RolePermissions' }
    Member:
      type: object
      properties:
        teamId:   { type: string, format: uuid }
        userId:   { type: string, format: uuid }
        roleId:   { type: string, format: uuid, nullable: true }
        joinedAt: { type: string, format: date-time }
    Invitation:
      type: object
      properties:
        id:        { type: string, format: uuid }
        teamId:    { type: string, format: uuid }
        email:     { type: string, format: email }
        status:    { type: string, enum: [pending, accepted, declined, cancelled] }
        invitedBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }

    # ── Git ───────────────────────────────────────────────────────────
...
security:
  - BearerAuth: []
paths:
...
  /api/teams:
    get:
      tags: [Teams]
      summary: Команды текущего пользователя (владелец или участник)
      responses:
        "200":
          description: Список команд
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Team' }
    post:
      tags: [Teams]
      summary: Создать команду
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "201":
          description: Созданная команда
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
  /api/teams/{id}:
    get:
      tags: [Teams]
      summary: Получить команду
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        "200":
          description: Команда
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
        "404": { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Teams]
      summary: Переименовать команду (только владелец)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "200":
          description: Обновлённая команда
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
    delete:
      tags: [Teams]
      summary: Удалить команду (только владелец)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        "204": { description: Удалено }
  /api/teams/{id}/roles:
    get:
      tags: [Teams]
      summary: Роли команды
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        "200":
          description: Роли
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Role' }
    post:
      tags: [Teams]
      summary: Создать роль (только владелец)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:        { type: string }
                permissions: { $ref: '#/components/schemas/RolePermissions' }
      responses:
        "201":
          description: Созданная роль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
  /api/teams/roles/{roleId}:
    patch:
      tags: [Teams]
      summary: Обновить роль (только владелец команды)
      parameters:
        - name: roleId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:        { type: string }
                permissions: { $ref: '#/components/schemas/RolePermissions' }
      responses:
        "200":
          description: Обновлённая роль
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
    delete:
      tags: [Teams]
      summary: Удалить роль
      parameters:
        - name: roleId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Удалено }
  /api/teams/{id}/members:
    get:
      tags: [Teams]
      summary: Участники команды
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        "200":
          description: Список участников
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Member' }
  /api/teams/{id}/members/{userId}:
    patch:
      tags: [Teams]
      summary: Назначить роль участнику (только владелец)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleId: { type: string, format: uuid, nullable: true }
      responses:
        "200":
          description: Обновлённый участник
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
    delete:
      tags: [Teams]
      summary: Исключить участника (владелец или сам участник)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Исключён }
  /api/teams/{id}/invitations:
    get:
      tags: [Invitations]
      summary: Список приглашений команды
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        "200":
          description: Приглашения
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Invitation' }
    post:
      tags: [Invitations]
      summary: Пригласить пользователя по email
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "201":
          description: Приглашение создано
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invitation' }

  # ═══════════════════════════════════════════════════════════════════
  # INVITATIONS
  # ═══════════════════════════════════════════════════════════════════
  /api/invitations:
    get:
      tags: [Invitations]
      summary: Ожидающие приглашения для email текущего пользователя
      parameters:
        - name: email
          in: query
          required: true
          schema: { type: string, format: email }
      responses:
        "200":
          description: Список ожидающих приглашений
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Invitation' }
  /api/invitations/{id}/accept:
    post:
      tags: [Invitations]
      summary: Принять приглашение в команду
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: email
          in: query
          required: true
          schema: { type: string, format: email }
          description: Email должен совпадать с email в приглашении
      responses:
        "204": { description: Приглашение принято, пользователь добавлен в команду }
        "403": { description: Email не совпадает с приглашением }
        "409": { description: Приглашение уже обработано }
  /api/invitations/{id}/decline:
    post:
      tags: [Invitations]
      summary: Отклонить приглашение
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: email
          in: query
          required: true
          schema: { type: string, format: email }
      responses:
        "204": { description: Приглашение отклонено }
  /api/invitations/{id}:
    delete:
      tags: [Invitations]
      summary: Отменить приглашение (только владелец команды)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        "204": { description: Приглашение отменено }

  # ═══════════════════════════════════════════════════════════════════
  # GIT
  # ═══════════════════════════════════════════════════════════════════
...